+++ steinerpy/algorithms/__init__.py
@@ -1,5 +1,5 @@
 # import all the algorithms here then we can do
 from .sstar import SstarHS, SstarBS, SstarMM, SstarMM0,\
                     SstarHSUN, SstarBSUN, SstarMMUN, SstarMM0UN,\
-                        SstarMMLP
+                        SstarMMLP, SstarHSLP, SstarMMUNLP
 from .kruskal import Kruskal